feat: display multiple treasuries on space page
